+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#26] Use correct label for dogu resources
+  - `dogu.name=name` is now valid
+
 ### Changed
 - [#24] Read dogu registry URL schema from cluster secret instead of config.